fix: use proper name convention when retrieving document contents
The SQL statement for retrieving a document and its contents was
returning the object property names in snake case, which is not the
JavaScript convention.

It has now been updated to return the names in camel case.  The
properties that fall under the Document model is also refactored to use
the variable that contains all of that.