chore: make add document-section route consistent
For the route that adds a document-section relationship, the variable
for document ID was called "docId".  To be consistent with the other
routes, it has now been changed to "documentId".

The JSON schema used has been changed to the urlParamsSchema, because
the document_x_sectionNewSchema was redundant.  urlParamsSchema
generalizes validation of the URL path parameter inputs and can be used
instead.  document_x_sectionNewSchema.json is deleted as a result.

The log for the section route code is updated to include the specific
username, documentId, and sectionId, because this is more helpful when
reading the logs.