feat: implement add/create experience
Created front-end and back-end functionality for creating a new
experience.

In the back-end, created the experiences route, service, and schema for
creating new experiences.  These were copied over from education.  As a
result, there is a lot of duplicate code, which will later have to be
refactored, but not now due to time.  Since it is the same code, tests
can be skipped.  Note for the experienceNew.json schema, max and min
lengths for properties are arbitrary and done to restrict ridiculous
text, but this decision may change in the future.

In the front-end, created the AddExperienceCard React component and
added it to the list of possible components to render, in the
SectionItemsList React component.  This is copied over from
AddEducationCard.

In the front-end, ResumeManagerApi has been updated to have
addExperience as a convenience method.

In the front-end, util/specificSectionsFuncs.js has been updated to use
ResumeManagerApi.addExperience.

Also installed missing nodemon package dependency.

Written tests have been skipped due to time, but basic manual tests
have been performed.