feat: implement adding text snippet to experience
Created front-end and back-end functionality for creating a new text
snippet for experiences.

* Back-end:
Implemented a new route in the experiences path to allow text snippets
to be added to experiences.  There is no new route dedicated to text
snippets, because text snippets can only exist within experiences,
skills, etc., like how experiences can only exist within documents.

Implemented the related service function, createTextSnippet, to do the
hard work.  It verifies both document and experience, so that users can
not access another user's data.  The document is required to be master,
because currently, only the master resume should be allowed to add,
edit, and delete info.

Generalization of the route and service code could have been done, but
there is a lot of code specific to education, and only the education
and experience sections are needed right now, so it isn't worth it.

A new JSON schema for new text snippets has been created.  Since the
data in text snippets is not specific to any section, this schema can
be used for all new text snippets.  The min length for content is
arbitrarily chosen; it is just there to have meaningful content.

* Front-end:
Created the AddTextSnippetCard React component to allow adding text
snippets to an experience or any other section item.  It is mostly
copied from AddEducationCard.

Created the TextSnippetsList React component to help separate out some
code.  It currently holds a placeholder React fragment for attaching
text snippets, in the rendered HTML.

In ExperienceCard, created a function called addTextSnippet to initiate
the API call and to update the document with a new text snippet.  This
method had to be done here, because adding a new text snippet required
the document and experience IDs.  Getting the document ID is easily
done through the existing document Object in the React context, but
getting the experience ID would require making another context within
ExperienceCard.  Since the addTextSnippet function is passed down as a
prop two times, using a context is skipped for now, but will be
considered in future refactoring.

Also for the addTextSnippet function in ExperienceCard, the document is
shallow copied, because it was easier to mutate the experience
argument/prop in ExperienceCard than to find the experience again in
the document.  This should not cause any issues, as the document will
be re-rendered, but unexpected behavior is something to watch out for.

Implemented the related ResumeManagerApi method, addTextSnippet, as a
convenience method for not having to write out the whole URL.

Written tests have been skipped due to time, but basic manual tests
have been performed.