fix: format timestamp in document contents return
When retrieving a document and its contents from the database in
resume-manager-backend/models/Document.getDocumentAndSectionContent,
the text snippet version was not being converted to a Date type, after
being received in the back-end app.  It stayed as a String type.
node-postgres only does type conversions for top-level data values.
This caused the text snippet version to have a format like
"2025-01-02T12:00:00-08:00", when a format like
"2025-01-02T20:00:00.000Z" was needed.  Because of the format, it was
not possible to use version values from the retrieved document and
contents.

The simplest solution was to modify the SQL statement in
getDocumentAndSectionContent and manually set the version to be a
String in the correct format.

Written tests have been skipped due to time, but basic manual tests
have been performed.