chore: update back-end common model tests
In resume-manager-backend\models\_testCommon.js, in the update suite,
the test for "Updates a ... with # properties" has had its parameterized
inputs for 0 and extra properties removed.  Since JSON schema is now
used, there will not be a case where a user can give no inputs or more
inputs than allowed.

Also changed lines involving "Object.keys(...).includes(...)" to use
the hasOwn method of Object.  This theoretically should be faster,
since an Array does not need to be created.