feat: implement edit document properties
Created back-end and front-end functionalities for editing a document's
properties, such as document name.

* Back-end:
In resume-manager-backend\routes\documents.js, renamed the path
parameter "docId" to "documentId" to be consistent with other route
paths and to clearly spell out what it is.  A local variable,
documentId, was destructured out of req.params, to clearly indicate the
inputs to the route at the top of the code.

* Front-end:
Created the EditDocumentForm React component to present a modal for
editing a document's properties.  This was copied and adapted from
NewDocumentForm.  Note that there are form inputs for setting a
document as a template and locking a document, but there are no
functionalities at all for those yet.  Consequently, these inputs may
be removed in the future.  The inputs exist as a reminder to implement
them.

In the Document React component, added EditDocumentForm and an edit
icon in the output for rendering.  The editDocument function was
implemented to handle submitting the updated properties to the relevant
method in ResumeManagerApi and to modify the documents and document
state.  Having the function in Document instead of in EditDocumentForm
helps keep the code logic in EditDocumentForm to be specific to
managing the form.

In the ContactInfoCard React component, renamed editIcon to pencilIcon,
to start consistency in having the SVG not solely be attributed to
editing.

Implemented updateDocument in ResumeManagerApi to help make it easier
to call the back-end API.

Written tests have been skipped due to time, but basic manual tests
have been performed.