feat: implement helper for nullifying Object values
Created a helper function that will replace an Object's empty String
values with nulls.  The original intention is to not have empty Strings
in database entries and to clearly indicate attributes that are not
supposed to contain anything, because having empty Strings for
attributes might make querying more difficult.  The Object passed-in is
supposed to be the properties used for updating database entries.