chore: rename experience_x_text_snippet model file
Renamed experience_x_text_snippet to experience_x_textSnippet in
resume-manager-backend\models\ to better suit JavaScript naming
convention, though it still does not fully follow the convention.  The
"_x_" in the name is still kept, because it more clearly indicates a
relationship between the left and right parts of the name than simply
an "X".  A decision between fully and not fully following the
convention still has not been finalized.