chore: remove redundant code in TextSnippet.update
In the back-end, the update method for the TextSnippet model was making
a database SELECT query in order to verify that the text snippet being
updated exists.  This was for cases of having stale text snippet
instances, from deleting a text snippet and then updating it.  This
extra database query is redundant, because the database schema has a
constraint for text snippet parent attribute values to reference
existing ones.  If a non-existent text snippet tries to update, it will
try to set the parent attribute to a non-existent version, causing the
database to throw a foreign key constraint violation.  As a result, the
redundant code has now been removed.

The AppServerError previously thrown by the redundant code is now moved
into the error callback function passed to PostgresDb.query.  It is
specifically thrown for PostgreSQL error code 23503 (foreign key
violation), because non-existent text snippets should only be the case
when that error code is thrown.  There are earlier checks for user
existence, which is the only other foreign key constraint.

resume-manager-backend\models\_testCommon.js has been updated
accordingly.  The extra arguments to create a new TextSnippet is now
needed, as those values are now being sent to the database in a SQL
INSERT statement before an Error is thrown.