feat: implement replacing text snippets in exp
Created a new method in the Experience_X_Text_Snippet model to update
all experiences_x_text_snippets for a particular text snippet with an
updated version.  This is supposed to be used alongside with updating a
text snippet, because documents do not automatically see updated text
snippets.  This is because updating a text snippet actually creates a
new one.  Using a SQL UPDATE in the replaceTextSnippet method helps
preserve the position of the text snippet.