feat: replace text snippet refs when updating
For the text snippet PATCH route for updating a text snippet, in the
service file textSnippetService.js, the function updateTextSnippet now
includes the method Experience_X_Text_Snippet.replaceTextSnippet.
replaceTextSnippet helps to update all references to the old snippet in
all documents, and experiences, to the new snippet, which would be the
expected behavior like when updating an education or experience.  This
extra step is necessary, because when text snippets update, a new text
snippet is actually created, which allows keeping the old text snippet
around.

Written tests have been skipped due to time, but basic manual tests
have been performed.