feat: implement reposition text snippet back-end
Implemented the route, service, and model functions to allow reordering
the text snippets in an experience in a document.

In resume-manager-backend\routes\experiences_x_textSnippets.js,
implemented the PUT route for reordering text snippets.

In resume-manager-backend\services\experienceXTextSnippetService.js,
created the updateExperienceXTextSnippetsPositions service function to
help validate document ownership and verify that all text snippet IDs
are given for an experience within a document.  All IDs are needed in
order to properly change the positions.  Note that the experience
ownership does not need to be validated, as mentioned in the code
comment.

In resume-manager-backend\models\experience_x_textSnippet.js, created
the updateAllPositions method to perform the database interaction.

In resume-manager-backend\models\textSnippet.js, created the
getAllForExperienceInDocument method to return the text snippets in
their new order.  getAllForExperience could not be modified to include
the new code, because the SQL statement is too different.

Written tests have been skipped due to time, but basic manual tests
have been performed.