docs: utilize dynamic names in logging
Changed hardcoded names used in logs to dynamic ones, so that they
don't need to be changed again if the class names change.

For class methods, the class name is dynamically retrieved.  For
instance methods, the class name and instance properties are retrieved
and logged.

Also in the Relationship model, the static delete method was
incorrectly trying to log properties from "this.constructor", which
only exists for instances.  For methods called on the class, the "this"
keyword already has all of the class properties.