test(back-end): added skipped TextSnippet tests
Added tests for getAllForExperience and getAllForExperienceInDocument
in TextSnippet model.

Also removed variables ClassRef and className, because the file is
specific to TextSnippet.  This should make reading the test file easier.

Also rewrote the beforeAll block for the TextSnippet describe block.
The raw SQL statement is replaced with a User.register call to make the
code shorter.  The then...catch chaining has been replaced with
async/await to make the code more readable.