docs(models): use correct return types in JSDocs
In the JSDocs of async functions, the assigned return data type was
incorrectly set as the unwrapped data type.  This has caused VSCode to
give a warning, "'await' has no effect on the type of this
expression.", on the await keyword when awaiting an asynchronous
function call.

All async functions in models have been corrected.