test: update test for documentNew JSON schema
When resume-manager-backend\schemas\documentNew.json was updated to
have the isTemplate property not be required, the associated unit tests
were not updated.

The unit test has now been updated by commenting the test case that
tests for a missing isTemplate.  It is commented to make it easier to
bring back that piece of code if template functionality gets
implemented.

Refs: 8815629