feat(api): remove document requirement for updating education
See reference commit's second paragraph for removal reason.  Basically,
the API route should not be involved in dealing with documents when
updating an education is independent of document.

In resume-manager-backend\routes\educations.js, the PATCH route for
updating an education has been updated with all references to document
removed.  In the associated test file, tests related to documents are
removed and the new URL is temporarily created in the describe block.

In resume-manager-backend\services\educationService.js, in the
updateEducation function, references to document have been removed.
Since updateEducation now does not contain logic, unit tests for it are
removed.

Refs: 5c2e12c