chore: add .gitignore after splitting repo
In order to have separate Git repositories for deployment, the original
repository had to push subtrees into new repositories inside the
back-end and front-end directories.  Since the back-end and front-end
directories are their own repositories now, they need their own
.gitignore files.  The original .gitignore file is in the parent
directory, and copies of it have been created in the back-end and
front-end.